+++ vulp/actuation/BulletInterface.cpp
@@ -42,12 +42,9 @@ BulletInterface::BulletInterface(const ServoLayout& layout,
     bullet_.setGravity(btVector3(0, 0, -9.81));
   }
   bullet_.setRealTimeSimulation(false);  // making sure
-  if (params.floor) {
-    bullet_.loadURDF(find_plane_urdf(params.argv0));
-  }
 
   // Load robot model
-  robot_ = bullet_.loadURDF(params.urdf_path);
+  robot_ = bullet_.loadURDF(params.robot_urdf_path);
   imu_link_index_ = find_link_index(bullet_, robot_, "imu");
   if (imu_link_index_ < 0) {
     throw std::runtime_error("Robot does not have a link named \"imu\"");