+++ CHANGELOG.md
@@ -6,14 +6,16 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- BulletInterface: Add a new parameter to allow environment URDFs.
+- BulletInterface: Add a new parameter `env_urdf_paths` to allow environment URDFs.
 
 ### Changed
 
 - Move exceptions to a dedicated namespace
+- BulletInterface: Rename `urdf_path` to `robot_urdf_path`.
 
 ### Removed
 
+- BulletInterface: Removed hardcoded ground plane URDF
 - Bazel: dependencies based on ``pip_parse`` from ``rules_python``
 
 ## [2.1.0] - 2024-01-03